refactor(web): lazy-load savings rate card to keep dashboard route chunk within perf budget (#2162)
